Refactored code with minor optimization
This commit creates the concept of 'data pipes' which consist of a
lockable vector of data, an atomic 'ready' flag, and a condition
variable. A regular pipe contains two 'half pipes' - upon completion
of a calculation we can now mark that individual pipe as complete
which will allow downstream computations to begin without waiting
for ALL pipes to be complete.